requirements.txt
customtkinter
ollama
numpy
pypdf
python-docx
packaging
pillow